requirements-dev.txt
-r requirements.txt

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-mock==3.12.0

# Security Testing
bandit==1.7.5
safety==3.0.1
semgrep==1.45.0

# Code Quality
pylint==3.0.3
black==23.12.1
flake8==7.0.0
mypy==1.7.1

# Development
ipython==8.18.1
ipdb==0.13.13
